Fix process tree design. Only passive mode so far.
The acceptor process always starts up with the application.
Passive startup is assumed for now. Upon receipt of a connection,
then, the acceptor will create an FSM process and send it a start
event. The server process is created by the FSM to handle messages
to and from the peer.

When active establishment is supported, the only necessary change
is to make the application spawn an FSM on startup, which will then
proceed to connect to the peer. Code for this is already in the FSM
but not in the application yet.