Use the OTP machinery to detect collisions.
Switch the FSM supervisor restart strategy to one_for_one. This
allows active sessions to be automatically started by the
supervisor itself. When a new TCP connection is established,
it is ignored if there's an FSM running for the connecting peer.
We start a new FSM if our side is configured as passive, or if an
active connection is not active (i.e. its FSM is not running).